Created base worker class. Move common functions and struct defines into it
