+++ src/c++/perf_analyzer/concurrency_manager.cc
@@ -134,7 +134,7 @@ ConcurrencyManager::ReconfigThreads(const size_t concurrent_request_count)
         start_sequence_id_, sequence_id_range_, on_sequence_model_, async_,
         max_concurrency_, using_json_data_, streaming_, shared_memory_type_,
         batch_size_, threads_config_, sequence_stat_, wake_signal_, wake_mutex_,
-        active_threads_, execute_, curr_seq_id_);
+        active_threads_, execute_, curr_seq_id_, distribution_);
 
     threads_.emplace_back(
         &ConcurrencyWorker::Infer, worker, threads_stat_.back(),
