+++ src/c++/perf_analyzer/concurrency_worker.h
@@ -30,7 +30,7 @@
 #include <random>
 
 #include "concurrency_manager.h"  // FIXME shouldn't need this
-#include "model_parser.h"
+#include "worker.h"
 
 namespace triton { namespace perfanalyzer {
 
@@ -42,7 +42,7 @@ namespace triton { namespace perfanalyzer {
 // at the same time. Thus it uses one single context as every infer context
 // creates a worker thread implicitly.
 //
-class ConcurrencyWorker {
+class ConcurrencyWorker : public Worker {
  public:
   ~ConcurrencyWorker();
   ConcurrencyWorker(ConcurrencyWorker&) = delete;
@@ -58,142 +58,44 @@ class ConcurrencyWorker {
       const SharedMemoryType shared_memory_type, const int32_t batch_size,
       std::vector<std::shared_ptr<ConcurrencyManager::ThreadConfig>>&
           threads_config,
-      std::vector<std::shared_ptr<LoadManager::SequenceStat>>& sequence_stat,
+      std::vector<std::shared_ptr<SequenceStat>>& sequence_stat,
       std::condition_variable& wake_signal, std::mutex& wake_mutex,
-      size_t& active_threads, bool& execute, std::atomic<uint64_t>& curr_seq_id)
-      : parser_(parser), data_loader_(data_loader), backend_kind_(backend_kind),
-        factory_(factory), sequence_length_(sequence_length),
-        start_sequence_id_(start_sequence_id),
-        sequence_id_range_(sequence_id_range),
-        on_sequence_model_(on_sequence_model), async_(async),
-        max_concurrency_(max_concurrency), using_json_data_(using_json_data),
-        streaming_(streaming), shared_memory_type_(shared_memory_type),
-        batch_size_(batch_size), threads_config_(threads_config),
-        sequence_stat_(sequence_stat), wake_signal_(wake_signal),
+      size_t& active_threads, bool& execute, std::atomic<uint64_t>& curr_seq_id,
+      std::uniform_int_distribution<uint64_t>& distribution)
+      : Worker(
+            parser, data_loader, sequence_stat, backend_kind,
+            shared_memory_type, batch_size, using_json_data, sequence_length,
+            start_sequence_id, sequence_id_range, curr_seq_id, distribution),
+        factory_(factory), on_sequence_model_(on_sequence_model), async_(async),
+        max_concurrency_(max_concurrency), streaming_(streaming),
+        threads_config_(threads_config), wake_signal_(wake_signal),
         wake_mutex_(wake_mutex), active_threads_(active_threads),
-        execute_(execute), curr_seq_id_(curr_seq_id)
+        execute_(execute)
   {
   }
 
   // FIXME What should be in constructor, and what here?
   void Infer(
-      std::shared_ptr<ConcurrencyManager::ThreadStat> thread_stat,
+      std::shared_ptr<ThreadStat> thread_stat,
       std::shared_ptr<ConcurrencyManager::ThreadConfig> thread_config);
 
  private:
   // FIXME -- should come into constructor? Also, same name as function
   // variables?
-  std::shared_ptr<DataLoader> data_loader_;
-  cb::BackendKind backend_kind_;  // FIXME changed from unique_ptr of backend_
   const std::shared_ptr<cb::ClientBackendFactory>& factory_;
-  const std::shared_ptr<ModelParser>& parser_;
 
-  // Map from shared memory key to its starting address and size
-  // FIXME -- does this need to be shared amongst all threads? Reference?
-  std::unordered_map<std::string, LoadManager::SharedMemoryData>
-      shared_memory_regions_;
-  std::default_random_engine rng_generator_;
-  std::uniform_int_distribution<uint64_t> distribution_;
-  // Current sequence id (for issuing new sequences)
-  std::atomic<uint64_t>&
-      curr_seq_id_;  // FIXME -- this needs to be shared across?
-  const uint64_t start_sequence_id_;
-  const uint64_t sequence_id_range_;
-  const size_t sequence_length_;
   const bool on_sequence_model_;
   const bool async_;
   const size_t max_concurrency_;
-  const bool using_json_data_;
   const bool streaming_;
-  const SharedMemoryType shared_memory_type_;
-  const int32_t batch_size_;
 
   // FIXME -- these need to be a reference? Should not be member variable?
   std::vector<std::shared_ptr<ConcurrencyManager::ThreadConfig>>&
       threads_config_;
-  std::vector<std::shared_ptr<LoadManager::SequenceStat>>& sequence_stat_;
   std::condition_variable& wake_signal_;
   std::mutex& wake_mutex_;
   size_t& active_threads_;
   bool& execute_;
-
-  // FIXME - all of these functions should be part of a shared base class!
-
-  /// Helper function to prepare the LoadManager::InferContext for sending
-  /// inference request. \param ctx The target LoadManager::InferContext object.
-  /// \return cb::Error object indicating success or failure.
-  cb::Error PrepareInfer(LoadManager::InferContext* ctx);
-
-  /// Helper function to prepare the LoadManager::InferContext for sending
-  /// inference request in shared memory. \param ctx The target
-  /// LoadManager::InferContext object. \return cb::Error object indicating
-  /// success or failure.
-  cb::Error PrepareSharedMemoryInfer(LoadManager::InferContext* ctx);
-
-  /// Updates the input data to use for inference request
-  /// \param inputs The vector of pointers to InferInput objects for all
-  /// possible inputs, potentially including optional inputs with no provided
-  /// data
-  /// \param valid_inputs The vector of pointers to InferInput objects to be
-  /// used for inference request.
-  /// \param stream_index The data stream to use for next data
-  /// \param step_index The step index to use for next data
-  /// \return cb::Error object indicating success or failure.
-  cb::Error UpdateInputs(
-      const std::vector<cb::InferInput*>& inputs,
-      std::vector<cb::InferInput*>& valid_inputs, int stream_index,
-      int step_index);
-
-  /// Updates the expected output data to use for inference request. Empty
-  /// vector will be returned if there is no expected output associated to the
-  /// step.
-  /// \param outputs The vector of outputs to get the expected data
-  /// \param stream_index The data stream to use for next data
-  /// \param step_index The step index to use for next data
-  /// \param data The vector of pointer and size of the expected outputs
-  /// \return cb::Error object indicating success or failure.
-  cb::Error UpdateValidationOutputs(
-      const std::vector<const cb::InferRequestedOutput*>& outputs,
-      int stream_index, int step_index,
-      std::vector<std::vector<std::pair<const uint8_t*, size_t>>>& data);
-
-  cb::Error ValidateOutputs(
-      const LoadManager::InferContext& ctx, const cb::InferResult* result_ptr);
-
-  void SetInferSequenceOptions(
-      const uint32_t seq_stat_index,
-      std::unique_ptr<cb::InferOptions>& options);
-  void InitNewSequence(int seq_stat_index);
-  uint64_t GetNextSeqId(int seq_stat_index);
-
-  /// Generate random sequence length based on 'offset_ratio' and
-  /// 'sequence_length_'. (1 +/- 'offset_ratio') * 'sequence_length_'
-  /// \param offset_ratio The offset ratio of the generated length
-  /// \return random sequence length
-  size_t GetRandomSequenceLength(double offset_ratio);
-
-  /// Helper function to update the inputs
-  /// \param inputs The vector of pointers to InferInput objects for all
-  /// possible inputs, potentially including optional inputs with no provided
-  /// data
-  /// \param valid_inputs The vector of pointers to InferInput objects to be
-  /// used for inference request.
-  /// \param stream_index The data stream to use for next data
-  /// \param step_index The step index to use for next data
-  /// \return cb::Error object indicating success or failure.
-  cb::Error SetInputs(
-      const std::vector<cb::InferInput*>& inputs,
-      std::vector<cb::InferInput*>& valid_inputs, const int stream_index,
-      const int step_index);
-
-  /// Helper function to update the shared memory inputs
-  /// \param inputs The vector of pointers to InferInput objects
-  /// \param stream_index The data stream to use for next data
-  /// \param step_index The step index to use for next data
-  /// \return cb::Error object indicating success or failure.
-  cb::Error SetInputsSharedMemory(
-      const std::vector<cb::InferInput*>& inputs, const int stream_index,
-      const int step_index);
 };
 
 }}  // namespace triton::perfanalyzer
